simplify CLI tool
